enterprise-ai-copilot/aiml/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.7.0
openai==1.35.0
langchain==0.2.0
langchain-openai==0.1.8
pandas==2.2.2
numpy==2.1.0
scikit-learn==1.5.0
python-dotenv==1.0.1
httpx==0.27.0
tiktoken==0.7.0
